Fixed some warnings when compiling with MSVC [patches:#31].
git-svn-id: https://svn.code.sf.net/p/open-cobol/code/branches/gnu-cobol-2.0@1045 ed166372-6744-4ac0-a67f-bb1ae9efa102

+++ libcob/move.c
@@ -917,7 +917,7 @@ cob_move_display_to_edited (cob_field *f1, cob_field *f2)
 
 		/* Put sign or currency symbol at the beginning */
 		if (sign_symbol || curr_symbol) {
-			if (floating_insertion) {			
+			if (floating_insertion) {
 				for (dst = end - 1; dst > f2->data; --dst) {
 					if (*dst == ' ') {
 						break;
@@ -1100,7 +1100,7 @@ indirect_move (void (*func) (cob_field *src, cob_field *dst),
 	cob_field	temp;
 	cob_field_attr	attr;
 
-	COB_ATTR_INIT (COB_TYPE_NUMERIC_DISPLAY, size, scale,
+	COB_ATTR_INIT (COB_TYPE_NUMERIC_DISPLAY, (unsigned short) size, (short) scale,
 			COB_FLAG_HAVE_SIGN, NULL);
 	temp.size = size;
 	temp.data = cob_malloc (size);
